Minor enhancements

+++ code/main.c
@@ -25,12 +25,12 @@
 
 #define TEST_SIZE 102400000
 int main(int argc, char** argv) {
-    int num_threads = sysconf(_SC_NPROCESSORS_ONLN);
-    int test_size = TEST_SIZE;
+    long num_threads = sysconf(_SC_NPROCESSORS_ONLN);
+    long test_size = TEST_SIZE;
     if (argc > 1)
-       test_size = atoi(argv[1]);
+       test_size = atol(argv[1]);
     if (argc > 2)
-       num_threads = atoi(argv[2]);
+       num_threads = atol(argv[2]);
     #ifdef UNIT_TESTS
     return DESTestCases();
     #else
